usb: core: phy: keep phy powered on if controller may wakeup
Some controllers may require the phy to be powered on to manage wakeup.
Moreover, some controllers enable wakeup only when there is a wakeup
enabled among its descendants.

So, modify usb_phy_roothub_suspend/resume with a new parameter which is the
number of enabled wakeup among the usb tree. And then power off the phy
only if the controller may not wakeup.

Change-Id: Ie9b34bf8b742074b6ce1f38ca0ac4d037878e3ea
Signed-off-by: Amelie Delaunay <amelie.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/187910
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@st.com>