usb: host: ehci-platform: fix wakeup management
device_wakeup_enable systematically returns -EINVAL due to the fact that
hcd->self.controller->power.can_wakeup is not initialized.
Check the device tree property wakeup-source, to know if the controller is
wakeup capable and set wakeup capable accordingly.

The controller will be able to wakeup the system only if a usb device with
wakeup enabled is plugged so, that's why we check both device_may_wakeup
or device_wakeup_path in suspend. If the conditions are met to wakeup the
system, then, enable the wake irq, otherwise let the system going into
deeper low power state.

Reverse things back on resume.

Change-Id: Id63d8b698a7df7fdae2010a5cd276506ed4c3a63
Signed-off-by: Amelie Delaunay <amelie.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/187911
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@st.com>