memory: stm32-omi: Add delay block support
One programmable delay is used for each OCTOSPI. These programmable
delays are used to tune the phase of:
  _ the RX sampling clock (or DQS) in order to sample the data in
    their valid window
  _ the TX launch clock in order to meet setup and hold constraints
    of TX signals versus the memory clock

Add delay block support used to tune the phase of Rx and Tx
The programmable delay block is based on a DLL that compensates
for voltage and temperature variations. The DLL features 32 taps
and covers the [50 MHz, 208 MHz] frequency range.

Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: I2690070d21ca6818f9dc915ec16a92c2b930a508