usb: host: ohci-platform: Add irq wakeup support
OHCI controller could be used as a wakeup source.

device_wakeup_enable systematically returns -EINVAL due to the fact that
hcd->self.controller->power.can_wakeup is not initialized.
Check the device tree property wakeup-source, to know if the controller is
wakeup capable and set wakeup capable accordingly.

The controller will be able to wakeup the system only if a usb device with
wakeup enabled is plugged so, that's why we check both device_may_wakeup
or device_wakeup_path in suspend. If the conditions are met to wakeup the
system, then, enable the wake irq, otherwise let the system going into
deeper low power state.

Reverse things back on resume.

Change-Id: Ibaebc265b1efb934b195b57d0d0c501ab029ac04
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/213950
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>