phy: stm32-usbphyc: add wakeup interrupt management per port
When a usbphyc phy is used by USBH controller, the wakeup signal is mapped
on EXTI for platform wakeup on STM32MP13.

The purpose of this patch is to get the optional wakeup interrupt per port
and enable the irq wake when phy is powered on, and disable the irq wake
in case the phy is powered off.

In case there is a device plugged on USBH, and wakeup is enabled, the phy
won't be powered off, so the wakeup interrupt will remain enabled.

Change-Id: I05a7d498b4245f68d011d5dcdd8ac8d03c06d9d9
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/213952
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>