usb: dwc2: don't disable/enable external vbus by system suspend/resume
It's not required to remove port power upon system suspend, as explained in
commit 113f86d0c302 ("usb: dwc2: Update partial power down entering by
system suspend")

Moreover, with current implementation runtime PM calls bus suspend/resume
routines, when no device is present on the port. In this case,
"b.port_connect_status" variable is 0, this makes:
- the _dwc2_hcd_suspend routine to skip power savings, vbus disable.
- the _dwc2_hcd_resume() is fully executed, including vbus enable
So, the vbus regulator enable count constantly increase. As a result, vbus
won't be turned off in case the role changes later to device mode.

So remove the disable/enable of external vbus in bus suspend/resume
routines to avoid this.

Fixes: 113f86d0c302 ("usb: dwc2: Update partial power down entering by system suspend")
Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: I28879a2f2d325ec2b6f02fd12e60b350cd0c75a0
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/229470
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>