usb: dwc2: handle suspended gadget when going to low power
In power_down mode 'none', in case the bus is suspended before going to
low power mode, dwc2_hsotg_suspend() will be a no-op. When resuming,
the controller registers may have been lost. But dwc2_hsotg_resume
will also be a no-op. So the gadget needs to be properly suspended, so it
can be later resumed.

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: I7deb8aecbec1420060ee05a74c4fa0a5e7315a7d
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/229475
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>