ARM: dts: stm32: add gc2145 entry within the stm32mp135f-dk
This adds the gc2145 node within the stm32mp135f-dk and link
the GC2145 to the ST_MIPID02 instead of the OV5640.
Remove the remote-endpoint property within the ov5640 node in
order to avoid a dtc warning due to non directional graph node
connection.
The property will be added at runtime by u-boot whenever the
camera is detected.

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Change-Id: I455619d2b3e2a35a449472cb5078cf3513cc013e