Revert "remoteproc: stm32: use workqueue to treat mailbox callback"
This reverts commit 714cf5e3846047052fae89c7b927b886cc69d60c.

For the rpmsg_i2c and rpmsg_intc drivers support we need to manage the
RPMsg in interrupt context.
This revert suppresses the creation of a workqueue, which
allowed to call the RPMsg framework in normal context.

Change-Id: Ia60c78ee416904b21504123e225d152daa3aea02
Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/286891
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/347948
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>