ARM: dts: stm32: adopt wakeup pin for stm32g0 typec in stm32mp135f-dk
USB Type-c controller (stm32g0) has an interrupt pin wired to a
STM32MP13 PWR wakeup pin on DK board. It can be used as a wakeup source,
to wakeup the system from standby mode, when changes are identified on the
Type-c connector.
All is configured in OPTEE, adopt it.

Change-Id: I23380dccb337665e06edc1c8d14caf58549efbc9
Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>