ARM: dts: stm32: use exti interrupt-map with stm32mp131
Stop using the internal table of the exti driver and use the more
flexible interrupt-map feature in DT.
By switching away from using the internal table, there is no need
anymore to use the specific compatible "st,stm32mp13-exti", which
was introduced to select the proper internal table.

Convert the driver table to interrupt-map property.
Add #address-cell=0 as the only child node that contains the irq
map is not located on a bus, and the irq map should not include
the 'child unit address'.
Switch the compatible string to the generic "st,stm32mp1-exti", in
place of the specific "st,stm32mp13-exti".

Older DT using compatible "st,stm32mp13-exti" will still work as
the driver is backward compatible, but there should not be around
any board using it, yet.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: I1a0615683e5dba5db5328aa96e38848f496851f3