dt-bindings: iio: adc: stm32-adc: add support for stm32mp25
Add stm32mp25 compatible for STM32MP25 SoC. It has 3 ADCs, split into two
blocks: ADC12 are tightly coupled (common registers). ADC3 is managed
independently.
Each ADC is a 12-bits successive approximation analog-to-digital converter,
with up to 20 multiplexed channels that can be configured as single ended
or differential.

STM32 ADC may be in a power domain which is the case for the STM32MP25
based boards. Allow a single 'power-domains' entry for STM32 ADC.

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: I8f7b114edaf7634014d24ab682f12d610f1b42fe