OPP: Provide old opp to config_clks on _set_opp
With the target opp, also pass the old opp to config_clks function.
This can be useful when a driver needs to take decision on what fequency
to set based on what is the current frequency without using a
clk_get_freq call.
Update the only user of custom config_clks (tegra30 devfreq driver) to
this new implementation.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>