ASoC: stm32: sai: temporary: manage dma burst mode on stm32mp25
The STM32 SAI peripheral does not support the DMA burst mode
on STM32MP25. Add a field in compatible structure to manage DMA
burst support capability.

This patch may be changed to removed dma burst mode support
on SAI for all SoCs. The status of burst mode utility has still
to be clarified.

Change-Id: I84447e8a4f0a81cdc7163653eca5a12e4f1a8474
Signed-off-by: Olivier Moysan <olivier.moysan@foss.st.com>