dmaengine: stm32-dma3: restore channel semaphore status after suspend
Depending on power state reached during suspend, CxSEMCR register could
have been reset, and semaphore released.
Try to take again the semaphore on resume. In case the semaphore can't be
taken, an error log with the channel number and channel user is displayed.

This requires the introduction of 2 new functions: stm32_dma3_pm_suspend(),
where the status of each channel is checked because we can't suspend if a
channel is still running; stm32_dma3_pm_resume(), where channel semaphore
is restored if it was taken before suspend.

Change-Id: I018d0f513e58cf7f67fc90dfd74c8e67311c7e32
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/370779
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>