block: add uImage.FIT subimage block driver
Add a small block driver which exposes filesystem sub-images contained
in U-Boot uImage.FIT images as block devices.

The uImage.FIT image has to be stored directly on a block device or
partition, MTD device or partition, or UBI volume.

The driver is intended for systems using the U-Boot bootloader and
uses the root device hint left by the bootloader (or the user) in
the 'chosen' section of the device-tree.

Example:
/dts-v1/;
/ {
        chosen {
                rootdisk = <&mmc0_part3>;
        };
};

Signed-off-by: Daniel Golle <daniel@makrotopia.org>