mxl-gpy: control LED reg from DT
Add dynamic configuration for the LED control registers on MXL PHYs.

This patch has been tested with MaxLinear GPY211C. It is unlikely to be
accepted upstream, as upstream plans on integrating their own framework
for handling these LEDs.

For the time being, use this hack to configure PHY driven device-LEDs to
show the correct state.

A possible alternative might be to expose the LEDs using the kernel LED
framework and bind it to the netdevice. This might also be upstreamable,
although it is a considerable extra amount of work.

Signed-off-by: David Bauer <mail@david-bauer.net>