net: ethernet: mediatek: fix ppe flow accounting for L2 flows
For L2 flows, the packet/byte counters should report the sum of the
counters of their subflows, both current and expired.
In order to make this work, change the way that accounting data is tracked.
Reset counters when a flow enters bind. Once it expires (or enters unbind),
store the last counter value in struct mtk_flow_entry.

Signed-off-by: Felix Fietkau <nbd@nbd.name>