arm64: dts: st: Add PCIe EP on the stm32mp257f-ev1 board
Configure PCIe EP and the ComboPHY with external pad clock on
MP25 eval board to be used as a device. Exclusive with pcie_rc
that is the default.
The EP is configured for CommonRef clock, need to deassert CLKREF#.
Note that the HW board needs to be modified to root the PCIE_CLK pins
to PCIE_CLKIN.

Signed-off-by: Christian Bruel <christian.bruel@foss.st.com>
Change-Id: I8acf6faa28921be9a70b5b3a0550a149e0b2b6ca