pinctrl: stm32: Fix reuse of irq mux lines
The driver programs a mux to connect the GPIO of a specific bank
to the parent interrupt controller.
The mux position is determined during irq_domain_ops::alloc(),
based on the content of struct irq_fwspec.
Once the GPIO is released, we need to also release the mux and
dispose the virq, so a following GPIO request will trigger a new
call to irq_domain_ops::alloc(). Currently this does not happen
and a new GPIO request fails. Running:
	gpiomon --chip 0 6
it requires the mux line #6 on position #0 for GPIO bank #0.
Quitting from 'gpiomon', the mux remains locked, and a following
call, e.g.:
	gpiomon --chip 2 6
that requires the same mux line on position #2, fails with message:
	stm32mp157-pinctrl soc:pinctrl@50002000: irq line 6 already requested.
	gpiomon: unable to request lines on chip /dev/gpiochip2: No such device or address

Call irq_dispose_mapping() in gpio_chip::free().

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: Ic25f4336742b004512df8b12eabe0a63e0b08683