arm64: dts: ti: k3-j7200: add reserved memory regions for lpm
Two memory regions are reserved for suspend to ram on j7200.
One is dedicated to TF-A, it uses this region to save its context during
suspend.
The second region is for R5 SPL, which uses it for its stacks and to store
some firmware images.
These regions are reserved from the linux point of view to avoid any
memory corruption.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>