arm64: dts: st: add rpmsg i2c stm32g0-typec device for stm32mp257f-dk board
Add the support of the stm32go typeC device on the rpmsg I2C adapter.
Instead of communicating with an external G0 processor, it communicates
with the coprocessor relying on a virtual i2c. G0  physical interrupts
are replaced by a rpmsg virtual interrupts controller through the
irq-rpmsg service.
notice that this impose that the rpmsg framework treat message under
interrupt context. this solution should be replaced by virtio-gpio and
virtio-i2c solutions.

Change-Id: I4249fdc0c38d40885fccd8d95f84b0fafde2af1b
Signed-off-by: Olivier BIDEAU <olivier.bideau@st.com>
Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>