arm64: dts: st: add rpmsg i2c stm32g0-typec device for stm32mp257f-ev1
Add the support of the stm32go typeC device on the rpmsg I2C adapter.
Instead of communicating with an external G0 processor, it communicates
with the coprocessor relying on a virtual i2c. G0  physical interrupts
are replaced by a rpmsg virtual interrupts controller through the
irq-rpmsg service.
notice that this impose that the rpmsg framework treat message under
interrupt context. this solution should be replaced by virtio-gpio and
virtio-i2c solutions.

Signed-off-by: Pankaj Dev <pankaj.dev@st.com>
Change-Id: Ide60b1de4b1a2b2bbfca859d108caad7cebfbe4d