remoteproc: stm32_rproc: Get resource table info from nvmem for the stm32mp25
For the stm32mp25 the nvmem framework is used to get resource table
information. for the stm32MP15 the syscon is used for compatibility
with the legacy
Use a function associated to the compatible to dispatch the method to
retrieve the information.

Change-Id: I769e4e52b2d6210fba12343c0279a7a9f62fc9f0
Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/326925
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/347954
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>