iio: adc: stm32-mdf: add iio support
Add IIO support to STM32 MDF.

NOT_UPSTREAMABLE

This version is not compatible with the SD modulator driver,
as the MDF IIO device is implemented as a channel provider.
(The SD modulator is currently a channel provider)
For time being the MDF device is not connected to a SD modulator
device, in the device tree.

Due to this restriction the scaling is not yet supported.
The MDF and SD modulator drivers have to adopt new IIO backend
framework to allow scaling support.

Signed-off-by: Olivier Moysan <olivier.moysan@foss.st.com>
Change-Id: I0f7d3bdb6941efdaf333f4932ad21a415b80521d
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/350254
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Domain-Review: Fabrice GASNIER <fabrice.gasnier@foss.st.com>