remoteproc: stm32: Add support of an OP-TEE TA to load the firmware
The new TEE remoteproc device is used to manage remote firmware in a
secure, trusted context. The 'st,stm32mp1-m4-tee' compatibility is
introduced to delegate the loading of the firmware to the trusted
execution context. In such cases, the firmware should be signed and
adhere to the image format defined by the TEE.

Change-Id: Iffeaa8b4ebd5a54ea5cf38dec9358140e35be065
Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>