dmaengine: stm32-dma3: clamp AXI burst using st,axi-max-burst-len
STM32 DMA3 can be interconnected with AXI3 or AXI4 busses. In case it is
interconnected with AXI3, the maximum burst length supported by AXI3
protocol is 16 beats, which is lower than the maximum burst length
supported by STM32 DMA3. So the programmed burst has to be shortened.

Change-Id: If0430ff300084258f3bc5ff8573b91afc8ca887a
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/370774
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>