dmaengine: stm32-dma3: avoid 64KB addressable space cross on LLI alloc
The documentation says "The user must program (...)_CxLLR for each LLIn to
not exceed the 64kB addressable space pointed by (...)_CxLBAR.

It means the <end> address of the LLI should not cross the 64KB addressable
space pointed by CxLBAR.

Use the boundary parameter for dma_pool, with SZ_64K.

For gen_pool, there is no boundary parameter, neither on pool creation nor
on allocation from pool. So the idea is to truncate the reserved-memory and
use the bigger part (the one before the 64KB, or the one after the 64KB).
In case of truncation a warning is displayed (reserved-memory is defined in
device tree and it may be updated after seeing this warning, to better
define the reserved-memory node).
Then change LLI gen_pool log, from dbg to info, to display the real base
address and size of the LLI gen_pool used.

Change-Id: I2f8469b4899e0067cc9e1c3b91a5250abcbe1822
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/370777
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>