arm64: dts: st: add HPDMA nodes on stm32mp211
The High Performance Direct Memory Access (HPDMA) controller is used to
perform programmable data transfers between memory-mapped peripherals
and memories (or between memories) via linked-lists.

There are 3 instances of HPDMA on stm32mp211, using stm32-dma3 driver, with
16 channels per instance and with one interrupt per channel.
Channels 0 to 7 are implemented with a FIFO of 32 bytes.
Channels 8 to 11 are implemented with a FIFO of 64 bytes.
Channels 12 to 14 are implemented with a FIFO of 128 bytes.
Channel 15 is implemented with a FIFO of 256 bytes.
Thanks to stm32-dma3 bindings, the user can ask for a channel with specific
FIFO size.

Change-Id: I693442cad09f7af0f92a4229a7b61035c55dde94
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>