EDITME: cover title for congatec-boad-controller
EDITME: describe the purpose of this series. The information you put
here will be used by the project maintainer to make a decision whether
your patches should be reviewed, and in what priority order. Please be
very detailed and link to any relevant discussions or sites that the
maintainer can review to better understand your proposed changes. If you
only have a single patch in your series, the contents of the cover
letter will be appended to the "under-the-cut" portion of the patch.

[1] https://git.congatec.com/x86/meta-congatec-x86/

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240503-congatec-board-controller-82c6b84cd4ea",
    "base-branch": "master",
    "prefixes": []
  }
}