Congatec Board Controller drivers
The Congatec Board Controller is a microcontroller embedded on the x86 SoM
of Congatec. It's able to manage lots of features, such as a watchdog, some
GPIOs, I2C busses ...

There is no datasheet or specific documentation for this Board Controller.
The only sources of information are the driver, library and tools provided
by Congatec in their yocto metalayer [1].

The Congatec implementation (available in [1]) doesn't follow the good
practice (a unique driver, and all accesses are done using custom ioctls).

This series implements an mfd driver, a gpio driver, a watchdog driver and
an I2C bus driver, to use the standard API from userspace.

For now, only the conga-SA7 module [2] is supported. For this board, the
Board Controller has:
- Two I2C busses
- 14 GPIOs
- A wathdog (with pretimeout support)

It also has temperature, voltage and fan sensors. They will be supported
later.

For the development, the conga-SEVAL board [3] was used.
With this board you have access to the 14 GPIOs, and the two I2C busses.
On each I2C bus, a 24c16 EEPROM is present by default.

To be able to drive GPIO 4, 5 and 6, a specific BIOS configuration is
needed: HD audio shall be disabled, and they shall be set in GPIO mode.

[1] https://git.congatec.com/x86/meta-congatec-x86/
[2] https://www.congatec.com/fileadmin/user_upload/Documents/Manual/SA70.pdf
[3] https://www.congatec.com/fileadmin/user_upload/Documents/Manual/SEVAL.pdf

To: Lee Jones <lee@kernel.org>
To: Linus Walleij <linus.walleij@linaro.org>
To: Bartosz Golaszewski <brgl@bgdev.pl>
To: Andi Shyti <andi.shyti@kernel.org>
To: Wim Van Sebroeck <wim@linux-watchdog.org>
To: Guenter Roeck <linux@roeck-us.net>
Cc: linux-kernel@vger.kernel.org
Cc: linux-gpio@vger.kernel.org
Cc: linux-i2c@vger.kernel.org
Cc: linux-watchdog@vger.kernel.org
Cc: thomas.petazzoni@bootlin.com
Cc: blake.vermeer@keysight.com
Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20240503-congatec-board-controller-v1-0-fec5236270e7@bootlin.com



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240503-congatec-board-controller-82c6b84cd4ea",
    "base-branch": "master",
    "prefixes": [],
    "history": {
      "v1": [
        "20240503-congatec-board-controller-v1-0-fec5236270e7@bootlin.com"
      ]
    }
  }
}