media: stm32: dcmipp: add TPG (test pattern generator) subdev
The dcmipp has an internal test-pattern-generator (TPG) allowing to send
patterns of various formats and sizes into the DCMIPP pipes.  Whenever
activated, it replaces the parallel or csi inputs of the dcmipp.

Since test-pattern can be of various formats/sizes/modes, proposal is
to expose this as a new subdev, in a very similar way as if it was a real
sensor.  It is thus connected to the sink pad of dcmipp_input and the sink
pad can either be connected to a real sensor (or bridge) or to the test
pattern generator.

Change-Id: Iaec20d8591f219dc002bcddab982b17e8fc11459
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/391850
Domain-Review: Philippe CORNU <philippe.cornu@foss.st.com>
Reviewed-by: Hugues FRUCHET <hugues.fruchet@st.com>
Reviewed-by: Philippe CORNU <philippe.cornu@foss.st.com>
Tested-by: Alain VOLMAT <alain.volmat@st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Alain VOLMAT <alain.volmat@st.com>