arm64: dts: st: Set soc@0 size-cell value to 2 to fix dma-ranges value on STM32MP21
"dma-ranges" is currently used on MP21 to describe the AXI bus hardware
translation for some AXI master peripherals: SDMMC[1-3], USBH, HPDMA.

This property is used to obtain a DMA address shifted by an offset of
-0x8000 0000. This shifted DMA address is then programmed in AXI
master's registers (of SDMMC[1-3], USBH, HPDMA peripherals).

As AXI translation is enabled on hardware side a shift of +0x8000 0000
is added by hardware for those masters.
This allow to access to the full 4GB DDR using AXI master 32bits address
registers (which is not possible STM32MP25/23).

Before this patch, if a buffer was allocated above the 4GB address space
(e.g. between 0x1 0000 0000 and 0x1 8000 000), it wasn't visible to the
DMA so the first mapping attempt fallback in SWIOTLB and the second
mapping was then possible as the bounce buffer used by the SWIOTLB is
allocated at the end of the 4GB (just before the 0x1 0000 0000) and so
visible for the DMA.

Change-Id: Id43d1885567e5258296741daed9f56ac969130c6
Signed-off-by: Clément Le Goffic <clement.legoffic@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/390368
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@foss.st.com>
Reviewed-by: Alain VOLMAT <alain.volmat@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Alexandre TORGUE <alexandre.torgue@foss.st.com>