Enable sensors support for the Congatec Board Controller
The Congatec Board Controller has some voltage, current, temperature
and fan sensors.
This series adds an hwmon driver to enable the support.

It depends on the series [1] (still under review), which adds initial
support of the Board Controller (including the core driver that
implements the communication with the Board Controller)

The sensors support has been tested on the conga-SA7 board with a
conga-SEVAL board.
The conga-SA7 board is the only board supported in [1].

[1] https://lore.kernel.org/all/20240503-congatec-board-controller-v2-0-681511a01c8f@bootlin.com/

Best Regards,

Thomas.

To: Jean Delvare <jdelvare@suse.com>
To: Guenter Roeck <linux@roeck-us.net>
To: Lee Jones <lee@kernel.org>
Cc: linux-kernel@vger.kernel.org
Cc: linux-hwmon@vger.kernel.org
Cc: thomas.petazzoni@bootlin.com
Cc: blake.vermeer@keysight.com
Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240809-congatec-board-controller-hwmon-e9e63d957d33",
    "prefixes": [],
    "prerequisites": [
      "change-id: 20240503-congatec-board-controller-82c6b84cd4ea:v3",
      "base-commit: v6.12-rc1"
    ]
  }
}