irqchip: stm32mp-exti1: Add support of genpd topology
Add a PM domain to manage the wake-up path for EXTI user,
this domain with GENPD_FLAG_ACTIVE_WAKEUP is used to manage EXTI1
dependency for the S2Idle / freeze with PSCI topology on STM32MP25 SoC.

A new IRQ power domain is needed as GENPD_FLAG_ACTIVE_WAKEUP is not managed
by the PSCI generic power domain RET_PD and in this family, the LPLV-Stop1
mode is not allowed when the EXTI1 is used for wake-up path.

The EXTI1 device is the only device in this new IRQ power domain, with the
parent domain found in device tree = RET_PD, and the wake patch for S2IDLE
is indicated with call of device_set_wakeup_path() in the suspend callback.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Iee92336da019b5c4c557f59f7133063846ceb28b
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/393892
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>