media: dcmipp: pixelcap: consider frame_skip when capturing frames
When frame skip is configured via postproc subdev pad frame_intervals,
there will be more VSYNCS than FRAMEEND events.  Capture mechanism is
based on the operations programmed into the dcmipp on each VSYNCS.
When frames are skipped, this means that some settings done on VSYNCS
are useless (since nothing will be generated after that VSYNC) and
moreover will lead to issue regarding the last frame captured before
reaching lack of buffer.  Indeed, when stopping the pipeline due to
a lack of buffer, the CPTREQ bit is cleared on the VSYNC in order to
have the pipeline stop after a last capture starting on the following
VSYNC.  Due to frame skip, this would end up in a situation of getting
the pipeline stopped while no frame has yet be generated since it has
been skipped.

This can be seen using YAVTA application which is specific in the sense
that, when using this application, the driver will reach the state of
having a single buffer queued prior to capturing the last frame.

In order to correct this, and moreover avoid unnecessary programming,
the capture device now keep track of the frame skip information in order
to only program the dcmipp targeting the VSYNC prior to the actual
frame end generated instead of all VSYNCS.

Change-Id: Ibba2746da203165c3135e1bc31f7e56e5e808a03
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/406294
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Alain VOLMAT <alain.volmat@st.com>
Domain-Review: Philippe CORNU <philippe.cornu@foss.st.com>
Tested-by: Alain VOLMAT <alain.volmat@st.com>
Reviewed-by: Hugues FRUCHET <hugues.fruchet@st.com>
Reviewed-by: Philippe CORNU <philippe.cornu@foss.st.com>