Add support for the AAEON UP board FPGA
This is the revival of the AAEON UP boards FPGA.
The latest version was sent one year ago [1].

I started a new series, as I re-wrote all the drivers using the AAEON
DKMS driver [2] as reference. I also read all series sent to the ML to
collect the comments sent by the maintainers and reviewers (to not
reproduce the same errors).

The FPGA acts as a level shifter between the Intel SoC and the HAT interfaces.
The pinctrl driver shall manage FPGA pins in tandem with the corresponding
Intel SoC pins. In the older series [2], this was done using Intel pinctrl
code (included in the FPGA pinctrl driver).
Now this code was removed and gpiolib is used to drive the SoC pins.

The FPGA also drives some LEDs.

This series adds support for UP and UP Squared boards.

More informations (like the HAT mapping) are available on the UP community
wiki [3].

[1] https://lore.kernel.org/all/20231031015119.29756-1-larry.lai@yunjingtech.com/
[2] https://github.com/up-division/pinctrl-upboard
[3] https://github.com/up-board/up-community/wiki

To: Linus Walleij <linus.walleij@linaro.org>
To: Bartosz Golaszewski <brgl@bgdev.pl>
To: Lee Jones <lee@kernel.org>
To: Pavel Machek <pavel@ucw.cz>
Cc: linux-gpio@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Cc: linux-leds@vger.kernel.org
Cc: thomas.petazzoni@bootlin.com
Cc: DanieleCleri@aaeon.eu
Cc: GaryWang@aaeon.com.tw
Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240930-aaeon-up-board-pinctrl-support-98fa4a030490",
    "prefixes": []
  }
}