remoteproc: k3-r5: Fix kernel panic at resume
If a remoteproc was attached at suspend and off at resume, the kernel
would panic on NULL pointer dereference.
(This happens when the SoC is off in suspend)

This is because the start method have been nullified for attached
remoteprocs.

To prevent the kernel panic and load the remoteproc firmware at resume,
we have to populate the prepare/unprepare/start/stop functions in
k3_r5_rproc_configure_mode().

Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>