drm/stm: fix compile issue
It is a temporary fix. The compile issue comes from the fact that commit
6a508bd6b305 ("drm/stm: refactor probe sequence") was not applied (due to
merge conflicts). This commit refactors the probe sequence and introduces
the label 'err_suspend'.

The patch 950-0918-drm-stm-set-max-segment-size.patch which corresponds to
commit 7529e7646317 ("drm/stm: set max segment size"), uses this label
which cause a compile issue.

For now to fix the compile issue, this commit removes the use of the
err_suspend label. But all drm patches should be integrated correctly to
fix the build.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>