pinctrl: stm32: show 'NO ACCESS' in case the pin is not available
Depending on package, the range of a pin could not be found. The driver is
unable to find the pin configuration if no range is found.
To avoid having an empty information instead of the configuration after pin
name in debugfs, add "NO ACCESS" when range is not found.

Change-Id: I782346755f8d2d845d3aad7004389da49c90673e
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/445618
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>