media: stm32: dcmipp: add default histogram area settings in dcmipp-isp.c
Since histogram capture is enabled as soon as statistics are enabled, it is
necessary to setup a default histogram area otherwise this would lead to
a bad configuration error from the histogram block. Unfortunately it is
not possible to perform this at the time of start_streaming the statcap
since this would potentially overright a previously done configuration
by the isp-param output device. The same way it is not possible to do
that on the isp-param output device since it is not mandatory to use
in order to get statistics. For this reason perform that when configuring
the isp block and set a default configuration.

Change-Id: I553d77d66a4950ebf21d7a92eb285bcc7e5238fd
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/446850
Tested-by: Alain VOLMAT <alain.volmat@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Philippe CORNU <philippe.cornu@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Alain VOLMAT <alain.volmat@st.com>
Domain-Review: Philippe CORNU <philippe.cornu@foss.st.com>