mmc: core: sdio: fix S18A broken devices
BCM4373 device is returning S18A to 1 even if we are at 1,8V. Based on
SDIO specification, it should be 0. The framework is then trying to switch
the voltage when it should not. To avoid such issue, the proposal is to
switch the voltage if S18A is set to 1 and the voltage set to 3,3V.

Change-Id: I800592802cc11888aa26cbdf1d8cee2b17ddc4b1
Signed-off-by: Christophe Kerello <christophe.kerello@foss.st.com>
Signed-off-by: Christophe Roullier <christophe.roullier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/446913
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Christophe KERELLO <christophe.kerello@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Tested-by: Christophe KERELLO <christophe.kerello@st.com>
Reviewed-by: Christophe KERELLO <christophe.kerello@st.com>