arm64: dts: st: fix the SCMI regulator unit-address for STM32MP boards
The SCMI regulators in device tree use format regulator@unit-address
and unit-address must be equal to the reg value in hexadecimal format,
but It is not the case for regulator used in STM32MP boards.

This patch solves this issue, even if the issue have no functional impact
and if DTC detects this issue only for simple-bus (warning simple_bus_reg).

Change-Id: Ib0df5b6b1476bb4a97bfda9797cccdef333c1f79
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/456318
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>