drm/stm: lvds: remove clock enable/disable
This issue rarely occurs when requesting to enable/disable the LVDS
pixel clock. The LVDS driver enables/disables the reference and peripheral
clocks. This is not necessary, as the clock framework does this itself
(the reference clock is the parent of the LVDS pixel clock).

Change-Id: I4289ce9d3636e4b3fbeef32572af7ef8634c439c
Signed-off-by: Yannick Fertre <yannick.fertre@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/458856
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>