usb: dwc3: force dual-role mode for stm32mp2
Force dual-role mode for stm32, and disable host and gadget modes.
This patch is a workaround to be able to use dwc3 in dual-role mode.

In OpenWrt core, the package kmod-dwc3 forces host mode. The dwc3 driver
with dual-role mode can be selected builtin in the target/subtarget kernel
configuration. But in case of a build with ALL_KMODS option, the kmod-dwc3
package is built, so the kernel configuration is overrided and host mode is
selected.

define KernelPackage/usb-dwc3
  TITLE:=DWC3 USB controller driver
  KCONFIG:= \
        CONFIG_USB_DWC3 \
        CONFIG_USB_DWC3_HOST=y \
        CONFIG_USB_DWC3_GADGET=n \
        CONFIG_USB_DWC3_DUAL_ROLE=n \

Setting CONFIG_USB_DWC3_HOST, CONFIG_USB_DWC3_GADGET, and
CONFIG_USB_DWC3_DUAL_ROLE are not needed. The kernel automatically selects
the right configuration:
- CONFIG_USB_DWC3_HOST if (USB && !USB_GADGET)
- CONFIG_USB_DWC3_GADGET if (!USB && USB_GADGET)
- CONFIG_USB_DWC3_DUAL_ROLE if (USB && USB_GADGET)

Once it will be fixed upstream, this patch will be removed and the dwc3
driver will be selected as kernel module.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>