remoteproc: stm32_rproc: Make MCU reset and hold_boot optional
On the STM32MP2x series, the Cortex-M33 boots first, and in this scenario,
Linux does not have the appropriate access to the reset controls.

In this configuration, the remoteproc driver attaches to the Cortex-M33
but should not manage the resets.

This patch makes the hold_boot and mcu reset declaration in the device
tree optional.

Change-Id: I01e06a2a1b39508688666de3df160de287ef85d0
Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/475151
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: Olivier MOYSAN <olivier.moysan@foss.st.com>