arm64: dts: st: add I3C4 support on stm32mp257f-ev1
I3C4 is available on GPIO expansion connector and on mikroBus connector of
stm32mp257f-ev1 board.
On GPIO expansion connector and on mikroBus connector, it can use the same
pins as I2C8 (PZ3 and PZ4 configured with i3c4 pins_a pinctrl), but there
are external pull-ups on the board (it affects the performance of the I3C).
On mikroBus connector, it could also be configured with i3c4 pins_b pinctrl
but it uses the same pins as SPI8, without external pull-up on the board.

Change-Id: I6f9ec764e2a575ffb33492813c46d8dde5b22b68
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/478606
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Alain VOLMAT <alain.volmat@st.com>