counter: introduce new count modes for stm32 timers
STM32 timers can use a trigger input, that influences the way the counter
gets updated. This signal, e.g. external to the timer, a.k.a ETR (External
Trigger) can be used to synchronize the counter start, the counter reset
or even gate the clock (or combine these modes) based on the trigger input
edge or level.
In these modes, the counter still counts on its internal clock. So the
choice is made to add the description for these modes to "count_mode"
and "count_mode_available" attributes.

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: Icc858030a9e930a23d9acbecafeb95ffa9072e0d
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/487191
Reviewed-by: Olivier MOYSAN <olivier.moysan@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Tested-by: Olivier MOYSAN <olivier.moysan@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>