FROMLIST: firmware: ti_sci: add support for restoring IRQs during resume
Some DM-Firmware are not able to restore the IRQ context after a
suspend-resume. The IRQ_CONTEXT_LOST firmware capability has been
introduced to identify this characteristic. In this case the
responsibility is delegated to the ti_sci driver, which maintains an
internal list of all requested IRQs. This list is updated on each
set()/free() operation, and all IRQs are restored during the resume_noirq()
phase.

Reviewed-by: Dhruva Gole <d-gole@ti.com>
Link: https://lore.kernel.org/all/20260407-ti-sci-jacinto-s2r-restore-irq-v5-2-97b28f2d93f9@bootlin.com/
Signed-off-by: Thomas Richard (TI) <thomas.richard@bootlin.com>