arm64: defconfig: sanitize coresight config in fragment-02-defconfig-addons
In the ARM64 defconfig, most of the coresight configuration switches
are already enabled as module. To avoid overriding them to built-in,
remove the coresight configuration in the fragment-02-defconfig-addons
except CONFIG_CORESIGHT_SOURCE_ETM4X that is needed for the ETM
peripheral. But change it to module so that it is consistent with the
rest.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: I0012f22896e49d73e6130ad8ddb10a324b8d7794
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/503603
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>