usb: dwc2: drd: add wait loop, for a/b-session valid
The BSESVLD may not be set, when the drd (set_role) calls
dwc2_hsotg_core_connect. So add a wait loop, to ensure it is set.
In case dwc2_hsotg_core_connect gets called without it, the pull-up
will not raise, so the host can't enumerate the gadget.
Similarly, the ASESVLD may not be properly set once
dwc2_ovr_avalid() has completed. So, add a wait loop as well.

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: I23b010eb6569eff7e8589bb0a18583a73816b7ec
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/512992
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>