usb: dwc2: wa for isoc packet tx fifo flush at periodic frame boundary
If an isochronous packet is still being transmitted, when the end of
periodic frame interval elapses, using dma desc on dwc2 controller,
the device wrongly flushes the pending packets from the tx fifo.
This can lead to a huge packet is sent on the bus, leading to the host
finally reset the bus, and re-enumerate it.

Tuning the periodic frame interval limit from 80% to 95%, on the
affected cores (2.70a to 4.00a) makes the issue very unlikely to occur.

Fixes: 197ba5f406cc ("Move DWC2 driver out of staging")
CC: stable@vger.kernel.org
Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: I6f745f72c09e6abe6b49ef180b5a8264d1030726
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/513480
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>