usb: dwc2: fix gadget state after device disconnect
In case the gadget has been suspended by the host, and it gets
disconnected, also take care to reset the suspended_from state.
In this scenario, the controller can be put into low power after the
disconnect, but the drd layer gets notified, triggering a resume
interrupt, setting the state back to the suspended_from state
(typically configured).
So it needs to be reset also to not attached.

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: I03b29f0562c8631f84ca013ef6aad001cabafa18
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/522343
Reviewed-by: Christian BRUEL <christian.bruel@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>